1049 ReCAPTCHA score check fix

See merge request antiadblock/codeine!72

+++ src/Drivers/Security/CAPTCHA/ReCAPTCHA.php
@@ -43,7 +43,7 @@
 
         if (isset($Result['success']) && isset($Result['score']))
         {
-            if ($Result['success'] && intval($Result['score']) > 0.5)
+            if ($Result['success'] && floatval($Result['score']) > 0.5)
                 ;
             else
             {
